Use new quote_identifiers rule before annotate_types
