+++ sqlglot/optimizer/qualify.py
@@ -4,10 +4,16 @@ import typing as t
 
 from sqlglot import exp
 from sqlglot.dialects.dialect import DialectType
-from sqlglot.optimizer import qualify_columns
 from sqlglot.optimizer.isolate_table_selects import isolate_table_selects
 from sqlglot.optimizer.normalize_identifiers import normalize_identifiers
+from sqlglot.optimizer.qualify_columns import (
+    qualify_columns as qualify_columns_func,
+    validate_qualify_columns as validate_qualify_columns_func,
+)
 from sqlglot.optimizer.qualify_tables import qualify_tables
+from sqlglot.optimizer.quote_identifiers import (
+    quote_identifiers as quote_identifiers_func,
+)
 from sqlglot.schema import Schema, ensure_schema
 
 
@@ -20,6 +26,7 @@ def qualify(
     expand_alias_refs: bool = True,
     infer_schema: t.Optional[bool] = None,
     isolate_tables: bool = False,
+    qualify_columns: bool = True,
     validate_qualify_columns: bool = True,
     quote_identifiers: bool = True,
     identify: bool = True,
@@ -44,11 +51,13 @@ def qualify(
         expand_alias_refs: Whether or not to expand references to aliases.
         infer_schema: Whether or not to infer the schema if missing.
         isolate_tables: Whether or not to isolate table selects.
+        qualify_columns: Whether or not to qualify columns.
         validate_qualify_columns: Whether or not to validate columns.
         quote_identifiers: Whether or not to run the quote_identifiers step.
             This step is necessary to ensure correctness for case sensitive queries.
             But this flag is provided in case this step is performed at a later time.
         identify: If True, quote all identifiers, else only necessary ones.
+
     Returns:
         The qualified expression.
     """
@@ -59,19 +68,15 @@ def qualify(
     if isolate_tables:
         expression = isolate_table_selects(expression, schema=schema)
 
-    expression = qualify_columns.qualify_columns(
-        expression,
-        schema,
-        expand_alias_refs=expand_alias_refs,
-        infer_schema=infer_schema,
-    )
-
-    if quote_identifiers:
-        expression = expression.transform(
-            qualify_columns.quote_identifiers, dialect, identify, copy=False
+    if qualify_columns:
+        expression = qualify_columns_func(
+            expression, schema, expand_alias_refs=expand_alias_refs, infer_schema=infer_schema
         )
 
+    if quote_identifiers:
+        expression = quote_identifiers_func(expression, dialect, identify, copy=False)
+
     if validate_qualify_columns:
-        qualify_columns.validate_qualify_columns(expression)
+        validate_qualify_columns_func(expression)
 
     return expression
