+++ sqlglot/optimizer/qualify_columns.py
@@ -4,9 +4,8 @@ import itertools
 import typing as t
 
 from sqlglot import alias, exp
-from sqlglot.dialects.dialect import DialectType
 from sqlglot.errors import OptimizeError
-from sqlglot.helper import case_sensitive, seq_get
+from sqlglot.helper import seq_get
 from sqlglot.optimizer.scope import Scope, traverse_scope, walk_in_scope
 from sqlglot.schema import Schema, ensure_schema
 
@@ -412,21 +411,6 @@ def _qualify_outputs(scope):
         new_selections.append(selection)
 
     scope.expression.set("expressions", new_selections)
-
-
-def quote_identifiers(
-    expression: exp.Expression, dialect: DialectType, identify: bool
-) -> exp.Expression:
-    """Makes sure all identifiers that need to be quoted are quoted."""
-    if isinstance(expression, exp.Identifier):
-        name = expression.this
-        expression.set(
-            "quoted",
-            identify
-            or case_sensitive(name, dialect=dialect)
-            or not exp.SAFE_IDENTIFIER_RE.match(name),
-        )
-    return expression
 
 
 class Resolver:
